+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v12.1
 --------------------------
 - Added summary of total amount required for pending materials
+- Added list of materials per supplier
 - This release may include more changes, for full details please refer to the CHANGELOG